Run the Gated DeltaNet in Swift, bit-identical to the contract (DC-029)
sources/DatacenterEngine/GatedDeltaNet.swift implements the whole layer: the
projection, the depthwise causal conv, the per-head gates, the chunked delta
rule with its padding and sequential chunk scan, the gated norm, and the output
projection. GatedDeltaNetTests asserts golden BIT PATTERNS emitted from the
Python contract for a single-chunk case and for a 70-position, two-chunk case
-- the scan over chunks only runs in the second, which is where a padding or
state-threading error hides.

It passes under -Onone AND -O. That second one is a claim rather than a
formality: a compiler that contracted a*b + c into an FMA would break every
invariant in this project silently, so "optimisation does not disturb the
numerics" is now a test.

Two real defects came out of the cross-language check, neither by inspection:

- the gate and decay arrays were written once per POSITION instead of once per
  HEAD, so every head would have used the last head's gates -- wrong in a way
  that still produces plausible numbers;
- the contract itself carried a second silu with a different branch structure
  (single-branch where the shared one is two-branch), which surfaced as a
  1-ULP disagreement across eleven thousand elements. It is now imported from
  one place, because a stated function with two implementations is a function
  with none.

5 new Swift tests, 24 engine tests in total, all passing in both
configurations.

+++ tools/ordered_qwen35.py
@@ -54,7 +54,10 @@ def softplus(x: np.ndarray, beta: float = 1.0, threshold: float = 20.0) -> np.nd
     out = np.empty_like(x)
     large = scaled > np.float32(threshold)
     out[large] = x[large]
-    out[~large] = f32(np.log1p(np.exp(scaled[~large])) / np.float32(beta))
+    # The contract's transcendental rule: evaluated in double, rounded to Float. Stating
+    # it that way rather than calling float32 `log1p`/`exp` is what lets the Swift
+    # implementation reproduce these bits.
+    out[~large] = f32(np.log1p(np.exp(np.float64(scaled[~large]))) / np.float64(beta))
     return out
